requirements.txt
# Automatically generated by https://github.com/damnever/pigar.
behave==1.2.6
chromedriver-autoinstaller==0.4.1
PyYAML==6.0
selenium==4.10.0
webdriver-manager==3.8.6
configparser
behavex==2.0.1
Appium-Python-Client==2.11.1
click==8.1.5